test(outlet/clickhouse): update TestClose() to look at worker state
ClickHouse state may not work if tests run in parallel.
