+++ outlet/clickhouse/functional_test.go
@@ -258,68 +258,3 @@ func TestMultipleServers(t *testing.T) {
 	}
 	t.Fatal("w.Flush(): cannot trigger connect error")
 }
-
-func TestClose(t *testing.T) {
-	server, database := clickhousedb.SetupClickHouseDatabase(t)
-	r := reporter.NewMock(t)
-	sch := schema.NewMock(t)
-	ctx, cancel := context.WithTimeout(t.Context(), 5*time.Second)
-	defer cancel()
-	ctx = clickhousego.Context(ctx, clickhousego.WithSettings(clickhousego.Settings{
-		"allow_suspicious_low_cardinality_types": 1,
-	}))
-
-	// Pin the observing client to one connection so the only variable in the
-	// server-side connection count is the worker's own.
-	dbConf := clickhousedb.DefaultConfiguration()
-	dbConf.Servers = []string{server}
-	dbConf.Database = database
-	dbConf.MaxOpenConns = 1
-	chdb, err := clickhousedb.New(r, dbConf, clickhousedb.Dependencies{Daemon: daemon.NewMock(t)})
-	if err != nil {
-		t.Fatalf("clickhousedb.New() error:\n%+v", err)
-	}
-	helpers.StartStop(t, chdb)
-	ch, err := clickhouse.New(r, clickhouse.DefaultConfiguration(), clickhouse.Dependencies{
-		ClickHouse: chdb,
-		Schema:     sch,
-	})
-	if err != nil {
-		t.Fatalf("clickhouse.New() error:\n%+v", err)
-	}
-	helpers.StartStop(t, ch)
-
-	// A raw table lets a flush succeed and open a connection.
-	if err := chdb.Exec(ctx, fmt.Sprintf("CREATE TABLE flows_%s_raw (%s) ENGINE = Memory", sch.ClickHouseHash(),
-		sch.ClickHouseCreateTable(
-			schema.ClickHouseSkipGeneratedColumns,
-			schema.ClickHouseSkipAliasedColumns))); err != nil {
-		t.Fatalf("chdb.Exec() error:\n%+v", err)
-	}
-
-	tcpConnections := func() int64 {
-		var v int64
-		if err := chdb.QueryRow(ctx, "SELECT value FROM system.metrics WHERE metric = 'TCPConnection'").Scan(&v); err != nil {
-			t.Fatalf("chdb.QueryRow() error:\n%+v", err)
-		}
-		return v
-	}
-
-	// Send a batch so the worker opens a connection, then close it.
-	bf := sch.NewFlowMessage()
-	bf.AppendUint(schema.ColumnDstAS, 65000)
-	bf.Finalize()
-	w := ch.NewWorker(1, bf)
-	w.Flush(ctx)
-	connected := tcpConnections()
-
-	// The server drops the connection asynchronously, so wait for the count to fall.
-	w.Close()
-	deadline := time.Now().Add(3 * time.Second)
-	for tcpConnections() >= connected {
-		if time.Now().After(deadline) {
-			t.Fatalf("Close() did not release the connection (was %d)", connected)
-		}
-		time.Sleep(50 * time.Millisecond)
-	}
-}
